refactor(agentic-ai): extract response handling logic into a response handler
